Patch 1.20 - 1.20.2:( Add functionality to the parking setup so that robots can only snap to their own team-specific boxes. For example, blue bots can only snap to the blue slot, and red bots can only snap to the red slot. If artifacts are outtaked into the goal when the classifier is full, simply “overflow” them by placing them in the correct human player zone - just like removing one artifact from the classifier. When clicking the “place sample robots” button, the two closest to the goal should be loaded with a full 3 balls, 2 purple and 1 green. Also when setting up the bots could you make the bottom 2 bots move to stand directly on the middle of its respective hypotenuse of the small triangle at the bottom. The blue should stand on the middle of the rising hypotenuse on the left side of the small bottom triangle vs the red one should start on the middle of the falling hypotenuse. When clicking the “place artifacts” button, three artifacts should be lined up on the side of the human player zone closest to the wall For any references needed for the starting config setters, an image has been added as context that shows what it should look like In the human player zone box add a section where u can manually add in balls to fill it, it should be able to be detected and when emptying the classifier updates the new balls into either its correct white square or into the “tunnel” encapsulated by colored tape just above this box. This means that for example, if 3 spots in the box are filled manually when opening the classifier six balls should fill up the square and the extra 3 should go above it vertically inside the red rectangle. Both rectangles can hold up to six extra balls vertically.
The human player zone (white box) should fill up to 9 slots, and the secret tunnel (rectangle above the human player zone) should fill up to 6 slots. Currently both only fill up to six.
Instead of using two different buttons for set up artifact and set up robots, have one singular set up button that does both simultaneously (this button that sets up both should replace the 2 preexisting buttons). In this, include artifact setup that fills up all spike marks in the way it was done previously, 3 in each human player zone and 3 loaded in each robot that is near the goal. Reference image for how it should look in the end is linked as context.
Also swap the logic for magnetic attachment for the bots to the coloured squares - as of right now the blue bots magnetize to the red squares and red bots magnetise to blue squares please switch this by ensure that the initial setup position of the bot isnt switched nor is the colour changed only that blue bots can only magnetise to the blue square and red bots can magnetise red squares. In short simply flip the logic for the coloured bots with the coloured squares.
Also move over the overflow in the rectangle above the white box to be tangent to the colored line - this requires moving the balls closer to the wall on each side.

If all slots are filled up, the ball objects should simply be removed from the game entirely
Currently, the three balls that are placed in the human player zone during the field setup are not counted as part of the nine slots. They should count as the first three

)